Fixed tests, still have some resume bugs
