+++ __tests__/noop.test.ts
@@ -1,0 +1,3 @@
+test('nothing', () => {
+    expect(true).toBe(true)
+})
