+++ src/components/layouts/base/index.tsx
@@ -2,10 +2,10 @@ import { graphql, StaticQuery } from 'gatsby'
 import * as React from 'react'
 import 'typeface-roboto'
 import 'typeface-roboto-mono'
-import '../../css/reset.css'
-import '../../css/syntax.css'
-import Head from '../head'
-import Header from '../header'
+import '../../../css/reset.css'
+import '../../../css/syntax.css'
+import Head from '../../head'
+import Header from '../../header'
 import './index.module.css'
 
 export default ({ children }: any) => (
