+++ src/components/layouts/empty/index.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import 'typeface-roboto'
 import 'typeface-roboto-mono'
-import '../../css/reset.css'
-import '../../css/syntax.css'
-import './empty.module.css'
+import '../../../css/reset.css'
+import '../../../css/syntax.css'
+import './index.module.css'
 
 export default ({ children }: any) => <div>{children}</div>
