+++ src/components/layouts/index.ts
@@ -1,0 +1,2 @@
+export { default as BaseLayout } from './base'
+export { default as EmptyLayout } from './empty'
