+++ src/pages/404.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react'
 import Helmet from 'react-helmet'
-import Layout from '../components/layouts'
+import { BaseLayout } from '../components/layouts'
 
 export default (props: any) => (
-    <Layout>
+    <BaseLayout>
         <div>
             <Helmet title="Not Found" />
             <h1>NOT FOUND</h1>
             <p>You just hit a route that doesn&#39;t exist.</p>
         </div>
-    </Layout>
+    </BaseLayout>
 )
