+++ src/templates/index.ts
@@ -1,0 +1,1 @@
+export { default as BlogPostTemplate } from './blog-post'
